+++ CHANGELOG.md
@@ -8,16 +8,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.3.0] - 2025-04-06
+
+### Changed
+
+- [new-feature] Added async interface
+
 ## [0.2.0] - 2024-02-17
 
 ### Changed
 
-- [breaking-change] Use an unconnected socket as the argument and pass in the inital server address to fix #1
+- [breaking-change] Use an unconnected socket as the argument and pass in the initial server address to fix #1
 
 ## [0.1.0] - 2023-08-21
 
 Initial release
 
 <!-- next-url -->
-[Unreleased]: https://github.com/kiranshila/tftp_client/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/kiranshila/tftp_client/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/kiranshila/tftp_client/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/kiranshila/tftp_client/compare/v0.1.0...v0.2.0